# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Qubirex API",
    "version": "4.3.1-phase0.1",
    "description": "Generated from the Express routers by scripts/openapi.js. Do not edit by hand."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "cookieSession": []
    },
    {
      "bearer": []
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "qbx_session",
        "description": "httpOnly SameSite=Strict session cookie. State-changing requests also need X-CSRF-Token."
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Same session token; accepted until the frontend moves to cookies (docs/decisions.md D-002)."
      }
    },
...
    "responses": {
      "Error": {
        "description": "Error. New routes (/api/employer, /api/verify, /api/practical) answer {error: {code, message}}; legacy routes answer {error: string}.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "oneOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/.well-known/did.json": {
      "get": {
        "tags": [
          "verify"
        ],
        "operationId": "get__well_known_did_json",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  }
}